requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
requests==2.31.0
beautifulsoup4==4.12.0
lxml==4.9.0
selenium==4.10.0
sqlite3-python==1.0
APScheduler==3.10.0
python-dotenv==1.0.0
